Help button, removed dups in data, moved partitioning logic inside each node, handled edge cases

+++ src/App.tsx
@@ -19,6 +19,12 @@ import Button from '@mui/material/Button';
 import Dialog from '@mui/material/Dialog';
 import AddIcon from '@mui/icons-material/Add';
 
+import List from '@mui/material/List';
+import ListItem from '@mui/material/ListItem';
+import ListItemText from '@mui/material/ListItemText';
+
+import HelpIcon from '@mui/icons-material/Help';
+
 // TODO: Shuffle dataset, allow changing MAX_PARTITION_SIZE
 function App() {
 	const ndx = useRef(0);
@@ -30,6 +36,7 @@ function App() {
 	const [insertInterval, setInsertInterval] = useState('Once');
 	const [modalOpen, setModalOpen] = useState(false);
 	const [modalData, setModalData] = useState<any[]>([]);
+	const [helpOpen, setHelpOpen] = useState(false);
 
 	const rowOptions = ['1', '5', '10', '50'];
 	const intervalOptions = ['Once', '0.1s', '0.5s', '1s', '3s'];
@@ -39,6 +46,7 @@ function App() {
 
 	const resetApp = () => {
 		pause();
+		ndx.current = 0;
 		shardApp.current.reset();
 	}
 
@@ -54,7 +62,6 @@ function App() {
 			const ms = 1000 * parseFloat(insertInterval.substring(0, insertInterval.length - 1));
 			intervalID.current = setInterval(insertData, ms);
 			setIsRunning(true);
-			console.log(ms);
 		}
 	}
 
@@ -65,6 +72,11 @@ function App() {
 
 	const insertData = () => {
 		let rowsLeft = parseInt(rowsToInsert, 10);
+		if (ndx.current >= data.length) {
+			alert("No more data to insert");
+			return;
+		}
+
 		while (ndx.current < data.length && rowsLeft) {
 			shardApp.current.insert(data[ndx.current]);
 			ndx.current++;
@@ -85,6 +97,10 @@ function App() {
 	}
 
 	const deleteNode = (id: number) => {
+		if (appState.nodes.length === 1) {
+			alert("It's a good idea to have at least one node storing our data");
+			return;
+		}
 		shardApp.current.deleteNode(id);
 	}
 
@@ -93,13 +109,19 @@ function App() {
   return (
     <div className="app-container">
 			<Dialog open={modalOpen} onClose={() => setModalOpen(false)}>
-				{(modalData.length)
-					? modalData.map((d, n) => <p key={n}>{d.Name}</p>)
-					: "No data"}
+				{(modalData.length) ? <DataList data={modalData} /> : "No data"}
+			</Dialog>
+			<Dialog open={helpOpen} onClose={() => setHelpOpen(false)} maxWidth="xl">
+				<Help />
 			</Dialog>
 			<div className='heading'>
-				<h2>Partitioning with Dynamic Rebalancing</h2>
-				<h3>By Mahamadou Juwara</h3>
+				<div className="name">
+					<h2>Partitioning with Dynamic Rebalancing</h2>
+					<h3>By Mahamadou Juwara</h3>
+				</div>
+				<IconButton aria-label="Reset" onClick={() => {setHelpOpen(true); pause();}}>
+					<HelpIcon />
+				</IconButton>
 			</div>
 			<div className="app">
 				<div className="toolbar">
@@ -116,7 +138,7 @@ function App() {
 						id="rows-to-insert"
 						value={rowsToInsert}
 						label="Rows to Insert"	// TODO: More succinct way to say this
-						onChange={(e) => setRowsToInsert(e.target.value)}
+						onChange={(e) => {setRowsToInsert(e.target.value); pause();}}
 					>
 						{rowOptions.map(r => <MenuItem key={r} value={r}>{r}</MenuItem>)}
 					</TextField>
@@ -127,7 +149,7 @@ function App() {
 						id="insert-interval"
 						value={insertInterval}
 						label="Frequency"
-						onChange={(e) => setInsertInterval(e.target.value)}
+						onChange={(e) => {setInsertInterval(e.target.value); pause();}}
 					>
 						{intervalOptions.map(r => <MenuItem key={r} value={r}>{r}</MenuItem>)}
 					</TextField>
@@ -143,13 +165,15 @@ function App() {
 						{partitionSizeOptions.map(r => <MenuItem key={r} value={r}>{r}</MenuItem>)}
 					</TextField> */}
 				</div>
-				<div className="node-container">
+				<div className="nodes-container">
 				{appState.nodes.map((node: Node, ndx: number) => (
 					<div key={ndx} className="node">
-						<IconButton size='small' aria-label="Delete" onClick={() => deleteNode(node.id)}>
-							<ClearIcon />
-						</IconButton>
-						<h4>{`Node ${node.id}`}</h4>
+						<div className="node-header">
+							<h4>{`Node ${node.id}`}</h4>
+							<IconButton size='small' aria-label="Delete" onClick={() => deleteNode(node.id)}>
+								<ClearIcon />
+							</IconButton>
+						</div>
 						<Divider flexItem />
 						<div className="partition-grid">
 						{node.partitions.map((part: Partition, n: number) => (
@@ -161,7 +185,7 @@ function App() {
 					</div>
 					)
 				)}
-				<IconButton size="large" sx={{width: 150, height: 150}} onClick={addNode}>
+				<IconButton size="large" className="add-icon" onClick={addNode}>
 					<AddIcon />
 				</IconButton>
 				</div>
@@ -170,4 +194,59 @@ function App() {
   );
 }
 
+const DataList = ({data}: {data: any[]}) => {
+	return (
+		<List>
+			{data.map((d, n) => (
+				<ListItem key={n}>
+						<ListItemText primary={d.Name} />
+				</ListItem>)
+			)}
+		</List>
+	);
+}
+
+const Help = () => {
+	return (
+		<div className="help">
+			<h3>What is partitioning with dynamic rebalancing?</h3>
+			<Divider />
+			<p><strong>Partitioning</strong> is splitting up your data into groups.</p>
+			<p><strong>Dynamically rebalancing</strong> is to set a limit on the amount of data allowed in each group, splitting the partition if it gets too large and merging it with an adjacent partition if it gets too small.</p>
+			<p>You may want to do this if:</p>
+			<ul>
+				<li>Your data is too large to be stored on a single computer</li>
+				<li>To scale, balancing the load of queries across multiple servers</li>
+				<li>For redundancy, </li>
+			</ul>
+			<br />
+
+			<h3>Implementation:</h3>
+			<Divider />
+			<p>There are some number of nodes, each node is responsible for some number of partitions and each partition is responsible for some range of keys.</p>
+			<p>When a piece of data (in this case, cars) gets inserted, the key for that piece of data is determined via a hash function.</p>
+			<p>The key tells us which partition that piece of data will reside in.</p>
+			<p>When a partition exceeds a certain size, it will be split in half.</p>
+			<p>When a new node is introduced, the new node will take on partitions from existing nodes to balance the load.</p>
+			<p>When a node is removed, it's partitions will be spread across the remaining nodes.</p>
+			<br/>
+
+			<h3>Controls</h3>
+			<Divider />
+			<p>The toolbar contains a:</p>
+			<ul>
+				<li><strong>Reset button</strong> to reset the state of the app</li>
+				<li><strong>Play/Pause button</strong> continue/stop inserting data</li>
+				<li><strong>Rows to Insert</strong> for the number of rows to insert</li>
+				<li><strong>Frequency</strong> for how frequently to insert the specified number of rows</li>
+			</ul>
+			<p>Each node has:</p>
+			<ul>
+				<li>A name</li>
+				<li>An "X" to remove that node from the cluster</li>
+				<li>Buttons representing each partition stored on that node, click to view it's data</li>
+			</ul>
+		</div>
+	);
+}
 export default App;
